requirements.txt
pandas
streamlit
numpy
pillow
PyMuPDF==1.24.5
torch
duckdb
langchain
langchain-community
langgraph
langchain_core
chromadb
tabulate
sentence-transformers
transformers
sqlglot
groq
openpyxl
langchain-chroma
langchain-groq
langchain-huggingface
unstructured jq